+++ semcore/dropdown-menu/src/DropdownMenu.jsx
@@ -7,12 +7,13 @@ import ScrollAreaComponent from '@semcore/scroll-area';
 import uniqueIDEnhancement from '@semcore/utils/lib/uniqueID';
 import i18nEnhance from '@semcore/utils/lib/enhances/i18nEnhance';
 import { localizedMessages } from './translations/__intergalactic-dynamic-locales';
+import { hasFocusableIn } from '@semcore/utils/lib/use/useFocusLock';
 
 import scrollStyles from './styleScrollArea';
 import style from './style/dropdown-menu.shadow.css';
 
 const KEYS = ['ArrowDown', 'ArrowUp', 'Enter', ' '];
-const INTERACTION_TAGS = ['INPUT', 'TEXTAREA'];
+const INTERACTION_TAGS = ['INPUT', 'TEXTAREA', 'BUTTON'];
 
 class DropdownMenuRoot extends Component {
   static displayName = 'DropdownMenu';
@@ -26,6 +27,8 @@ class DropdownMenuRoot extends Component {
     i18n: localizedMessages,
     locale: 'en',
   };
+
+  popperRef = React.createRef();
 
   itemProps = [];
 
@@ -42,22 +45,45 @@ class DropdownMenuRoot extends Component {
 
   bindHandlerKeyDown = (place) => (e) => {
     const amount = e.shiftKey ? 5 : 1;
+    const targetTagName = e.target.tagName;
 
-    if (e.key === ' ' && INTERACTION_TAGS.includes(e.target.tagName)) return;
-    if (e.key === 'Enter' && e.target.tagName === 'TEXTAREA') return;
+    if (e.key === ' ' && INTERACTION_TAGS.includes(targetTagName)) return;
+    if (e.key === 'Enter' && (targetTagName === 'TEXTAREA' || targetTagName === 'BUTTON')) return;
+
+    const { visible, interaction } = this.asProps;
+    const element = this.popperRef.current;
+
+    if (
+      place === 'popper' &&
+      interaction !== 'focus' &&
+      visible &&
+      e.key === 'Tab' &&
+      hasFocusableIn(element)
+    ) {
+      this.handlers.highlightedIndex(null);
+
+      return;
+    }
+
     if (!KEYS.includes(e.key)) return;
 
     e.preventDefault();
 
+    const isVisible = this.asProps.visible;
+
     this.handlers.visible(true);
 
     switch (e.key) {
-      case 'ArrowDown':
-        this.moveHighlightedIndex(amount, e);
+      case 'ArrowDown': {
+        isVisible && this.moveHighlightedIndex(amount, e);
+        targetTagName === 'BUTTON' && element.focus();
         break;
-      case 'ArrowUp':
-        this.moveHighlightedIndex(-amount, e);
+      }
+      case 'ArrowUp': {
+        isVisible && this.moveHighlightedIndex(-amount, e);
+        targetTagName === 'BUTTON' && element.focus();
         break;
+      }
       case ' ':
       case 'Enter':
         if (this.highlightedItemRef.current) {
@@ -97,6 +123,7 @@ class DropdownMenuRoot extends Component {
     const { uid, disablePortal, ignorePortalsStacking } = this.asProps;
 
     return {
+      ref: this.popperRef,
       tabIndex: 0,
       onKeyDown: this.bindHandlerKeyDown('popper'),
       id: `igc-${uid}-popper`,