+++ semcore/pills/src/index.d.ts
@@ -11,6 +11,7 @@ export interface IPillsProps<T extends PillsValue = PillsValue>
   extends PillsProps<T>,
     UnknownProperties {}
 export type PillsProps<T extends PillsValue = PillsValue> = NeighborLocationProps &
+  KeyboardFocusProps &
   BoxProps & {
     /** Pills size */
     size?: 'l' | 'm';
@@ -32,7 +33,12 @@ export type PillsProps<T extends PillsValue = PillsValue> = NeighborLocationProp
      * while `tabs` behavior for navigation and layout.
      * @default tabs
      */
-    behavior?: 'tabs' | 'radio';
+    behavior?: /** @deprecated use `manual` */
+      | 'tabs'
+      /** @deprecated use `auto` */
+      | 'radio'
+      | 'auto'
+      | 'manual';
   };
 
 /** @deprecated */