+++ semcore/breakpoints/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [1.12.3] - 2023-11-24
+
+### Fixed
+
+- Matcher function in `MediaList` to use `defaultIndex`.
+
+## [1.12.2] - 2023-11-21
+
+### Changed
+
+- Version prepatch update due to children dependencies update (`@semcore/utils` [4.10.2 ~> 4.10.3]).
+
 ## [1.12.1] - 2023-11-09
 
 ### Changed