+++ semcore/color-picker/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [2.16.2] - 2023-11-21
+
+### Changed
+
+- Version prepatch update due to children dependencies update (`@semcore/button` [5.11.2 ~> 5.11.3], `@semcore/utils` [4.10.2 ~> 4.10.3]).
+
 ## [2.16.1] - 2023-11-14
 
 ### Changed