+++ semcore/core/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [2.10.0] - 2023-11-24
+
+### Changed
+
+- Types for `abstract class RootComponent`.
+
+## [2.9.2] - 2023-11-21
+
+### Changed
+
+- Version prepatch update due to children dependencies update (`@semcore/utils` [4.10.2 ~> 4.10.3]).
+
 ## [2.9.1] - 2023-11-09
 
 ### Changed