+++ semcore/data-table/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.17.2] - 2023-11-21
+
+### Changed
+
+- Version prepatch update due to children dependencies update (`@semcore/utils` [4.10.2 ~> 4.10.3], `@semcore/core` [2.9.1 ~> 2.9.2]).
+
+## [4.17.1-prerelease.1] - 2023-11-20
+
+### Changed
+
+- Removed animation from sort icon.
+
 ## [4.17.0] - 2023-11-13
 
 ### Changed