+++ semcore/input-number/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [5.13.0] - 2023-11-22
+
+### Added
+
+- Reexport utils and types.
+
+## [5.12.2] - 2023-11-21
+
+### Changed
+
+- Version prepatch update due to children dependencies update (`@semcore/input` [4.12.1 ~> 4.12.2], `@semcore/utils` [4.10.2 ~> 4.10.3], `@semcore/core` [2.9.1 ~> 2.9.2]).
+
 ## [5.12.1] - 2023-11-09
 
 ### Changed