+++ semcore/ui/CHANGELOG.md
@@ -1,3 +1,15 @@
+## [15.35.3] - 2023-11-21
+
+### @semcore/utils
+
+- **Changed** Values for `--facebook`, `--twitter` and `--linkedIn` variables to new ones.
+
+## [15.35.2] - 2023-11-20
+
+### @semcore/data-table
+
+- **Changed** Removed animation from sort icon.
+
 ## [15.35.1] - 2023-11-14
 
 ### @semcore/button