+++ semcore/utils/CHANGELOG.md
@@ -2,6 +2,30 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.13.0] - 2023-11-24
+
+### Added
+
+- Types for enhances.
+
+## [4.12.0] - 2023-11-23
+
+### Added
+
+- Pass `props` from component into `onNeighborChange` in `a11yEnhance` as a second args.
+
+## [4.11.0] - 2023-11-22
+
+### Added
+
+- New `findAllComponents` (for passed `Children`) util method.
+
+## [4.10.3] - 2023-11-21
+
+### Changed
+
+- Values for `--facebook`, `--twitter` and `--linkedIn` variables to new ones.
+
 ## [4.10.2] - 2023-11-09
 
 ### Fixed